9/9 cont'd

+++ include/autonomous.h
@@ -3,7 +3,7 @@
 
 void SetTimeout(int);
 void moveForward(int dist, int speed, int seconds); //(int dist, int speed, int timeout(seconds))//
-void TurninPlace(int dist, int speed, int seconds); //(int dist, int speed, int timeout(seconds))//
+void TurninPlace(double dist, int speed, int seconds); //(int dist, int speed, int timeout(seconds))//
